+++ iscc_sdk/__init__.py
@@ -29,3 +29,4 @@ from iscc_sdk.pdf import *
 from iscc_sdk.epub import *
 from iscc_sdk.docx_ import *
 from iscc_sdk.thumbnail import *
+from iscc_sdk.utils import *